fix: mudança e adição de anotações da classe de serviço, mudança de um dos campos da entidade e adição da dependencia do sql no pom.xml

+++ src/main/java/org/example/restapi/entity/Produto.java
@@ -12,5 +12,5 @@ public class Produto {
     private Long id;
     private String nome;
     private double preco;
-    private double quantidade;
+    private int quantidade;
 }
